Rename d3nlib to bandlib
